requirements.txt
streamlit
pandas
requests
python-dotenv
xlsxwriter
fpdf